Refactor: Guestbook service 패키지 생성
